pyproject.toml
[project]
name = "telegram-agent-mcp-client"
version = "0.10.3"
description = "Telegram Agent MCP Client"
authors = [{ name = "philogicae" }]
requires-python = ">=3.11,<4.0"
readme = "README.md"
license = "MIT"
dependencies = [
    "rich",
    "pydantic",
    "addict",
    "python-dotenv",
    "pyjson5",
    "langchain",
    "langchain-google-genai",
    "langchain-ollama",
    "langchain-core",
    "langchain-mcp-adapters",
    "langgraph",
    "pytelegrambotapi",
    "aiohttp",
    "aiosqlite",
    "langgraph-checkpoint-sqlite",
    "nest-asyncio",
    "rqbit-mcp",
    "telegramify-markdown",
    "langgraph-swarm",
    "graphiti-core[google-genai]",
    "unidecode",
]

[project.urls]
Repository = "https://github.com/philogicae/telegram-agent-mcp-client"
Release = "https://github.com/philogicae/telegram-agent-mcp-client/releases"
DeepWiki = "https://deepwiki.com/philogicae/telegram-agent-mcp-client"

[project.scripts]
telegram-agent-mcp-client = "telegram_agent.__main__:cli"

[dependency-groups]
dev = ["ruff", "pytest", "pytest-asyncio", "mypy"]

[tool.hatch.build]
include = ["telegram_agent"]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
